fix: delete validation for File
