+++ src/services/achievements/achievements.service.ts
@@ -7,7 +7,7 @@ import {
   PayloadAssignAchievement,
 } from '@/models/schema';
 import { API_PAYLOAD, API_RESPONSE, API_SINGLE_RESPONSE } from '@/models/types';
-import { axiosErrorHandler, ParseToFormData } from '@/utils/axios';
+import { axiosErrorHandler, ParseToFormData } from '@/utils';
 import { valuesWithData } from '@/utils/objects';
 
 export const ListAchievementService = async (
