+++ src/utils/axios.ts
@@ -1,8 +1,5 @@
 import { API_RESPONSE, I_VALIDATION_ERRORS } from '@/models/types';
-import { UploadFile } from 'antd';
 import axios, { AxiosError } from 'axios';
-
-const INPUT_FILE_KEY = 'imageFile';
 
 export interface CustomError extends Error {
   statusCode?: number;
@@ -29,43 +26,4 @@ export const axiosErrorHandler = (
   }
 
   throw err;
-};
-
-export const ParseToFormData = (payload: Record<string, unknown>): FormData => {
-  try {
-    const form = new FormData();
-
-    Object.entries(payload).forEach(([key, value]) => {
-      if (value === undefined || value === null) {
-        return;
-      }
-
-      if (key === INPUT_FILE_KEY) {
-        const fileList = value as UploadFile[];
-        const firstFile = fileList[0];
-        if (firstFile?.originFileObj) {
-          form.append(key, firstFile.originFileObj as Blob);
-        }
-        return;
-      }
-
-      if (Array.isArray(value)) {
-        value.forEach((item) => {
-          form.append(`${String(key)}[]`, item.toString());
-        });
-        return;
-      }
-
-      if (value instanceof Blob || value instanceof File) {
-        form.append(String(key), value);
-        return;
-      }
-
-      form.append(String(key), value.toString());
-    });
-
-    return form;
-  } catch (error) {
-    throw error;
-  }
 };
